% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.r
\name{get_gender}
\alias{get_gender}
\title{Request gender predictions from NamSor}
\usage{
get_gender(given, family, api_key)
}
\arguments{
\item{given}{A vector of given names (i.e. first names).}

\item{family}{A vector of family names (i.e. surnames or last names).}

\item{api_key}{Your NamSor API key. Get one at \url{https://namsor.app/}}
}
\value{
A data frame with the input names, API URL, predicted gender and scale.
}
\description{
Sends given and family names to NamSor and returns its gender prediction and scale.
}
\examples{
# Prepare input vectors
first_name <- c("Volha", "Charles", "Donald")
last_name <- c("Chykina", "Crabtree", "Duck")

# Expected output columns
expected_cols <- c("id", "first_name", "last_name", "api_url", "scale", "gender")
print(expected_cols)

\dontrun{
# Note: the vectors of first and last names should be the same length.
key <- "YOUR_NAMSOR_API_KEY"
y <- get_gender(first_name, last_name, key)
y
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
