% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicites.R
\name{get_ethnicities}
\alias{get_ethnicities}
\title{Classifies names based on 6 U.S. ethnicities}
\usage{
get_ethnicities(x)
}
\arguments{
\item{x}{A vector of names.}
}
\value{
An object that contains that classifies any inputted names(s) accoridng to 6 different U.S. ethnicities.
}
\description{
Returns an object that classifies any inputted name(s) according to 6 different U.S. ethnicities.
}
\examples{
x <- c("Charles Crabtree", "Volha Chykina")
get_ethnicities(x)
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
