% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalities.r
\name{get_nationalities}
\alias{get_nationalities}
\title{Classifies names based on 39 leaf nationalities}
\usage{
get_nationalities(x)
}
\arguments{
\item{x}{A vector of names.}
}
\value{
An object that classifies inputted names according inputted name belongs to 39 different leaf nationalities.
}
\description{
Returns an object that classifies inputted names according to 39 different leaf nationalities.
}
\examples{
x <- c("Charles Crabtree", "Volha Chykina")
get_nationalities(x)
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
